Created AbstractJoin and added innerJoin
