+++ src/Query/InnerJoin.php
@@ -5,13 +5,13 @@ namespace Happyr\DoctrineSpecification\Query;
 use Doctrine\ORM\Query;
 use Doctrine\ORM\QueryBuilder;
 
-class LeftJoin extends AbstractJoin
+class InnerJoin extends AbstractJoin
 {
     /**
      * {@inheritdoc}
      */
     protected function getJoinType()
     {
-        return 'leftJoin';
+        return 'innerJoin';
     }
 }
