+++ src/Query/Join.php
@@ -2,46 +2,16 @@
 
 namespace Happyr\DoctrineSpecification\Query;
 
+use Doctrine\ORM\Query;
 use Doctrine\ORM\QueryBuilder;
 
-/**
- * Class Join.
- */
-class Join implements QueryModifier
+class Join extends AbstractJoin
 {
     /**
-     * @var string field
+     * {@inheritdoc}
      */
-    private $field;
-
-    /**
-     * @var string alias
-     */
-    private $newAlias;
-    private $dqlAlias;
-
-    /**
-     * @param string $field
-     * @param string $newAlias
-     * @param string $dqlAlias
-     */
-    public function __construct($field, $newAlias, $dqlAlias = null)
+    protected function getJoinType()
     {
-        $this->field = $field;
-        $this->newAlias = $newAlias;
-        $this->dqlAlias = $dqlAlias;
-    }
-
-    /**
-     * @param QueryBuilder $qb
-     * @param string       $dqlAlias
-     */
-    public function modify(QueryBuilder $qb, $dqlAlias)
-    {
-        if ($this->dqlAlias !== null) {
-            $dqlAlias = $this->dqlAlias;
-        }
-
-        $qb->join(sprintf('%s.%s', $dqlAlias, $this->field), $this->newAlias);
+        return 'join';
     }
 }
